datapunk/datapunk-stream/requirements.txt
# API and Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Task Processing
celery>=5.3.0
temporal-python>=1.0.0
prefect>=2.14.0

# Message Brokers and Caching
redis>=5.0.1
rabbitmq-client>=3.8.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Monitoring
prometheus-client>=0.19.0

# PII Detection
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

